Fix GateAction parsing in 2015 07
I'm not entirely sure what I was thinking here, since s.parse can never
fail in this particular case, and even if it could there would be no
reason to override the error message.

I guess the 2024 edition worked out I was an idiot? But whatever, fixing
this is somehow mandatory for the edition update